WIP
This was a try to extend the dynamic and static walk for their remove block related
duties. I would recommend starting this again in the future and doing all the copy
and pasting in one go since it's otherwise quite hard to keep track of what's missing.

One concern I had so far was if the removed blocks source is a component or a component instance address.
For components their expansion changes how they are addressed since they are expanded.
I thought removed blocks are different (and they might be). A removed block should always refer to a
component instance, with or without expansion since the source of its address is related to the source
attribute rather than the name label or similar things.

I wonder if we need to skip the static walk step since if we have an expansion we can not construct the right
instance value without evaluating for_each. So maybe we don't do a static walk and go straight to the dynamic one.