skip unneeded variable validation during destroy
Arbitrary expressions cannot be evaluated during destroy, because the
"state" of the state is unknown. Missing resources or invalid data will
cause evaluations to fail, preventing any progress in the destroy
operations.

The way we handle this with providers and temporary values that need to
be re-evaluated during destroy, is that they are selectively pruned out
of the graph during the destroy process. Now that variable validations
can reference other values in the configuration, which added another
evaluation node to the graph, that new node must also act like it's a
temporary value akin to variables, locals, and outputs.

This has some implications for validations in general, since they were
always run during a destroy. However now that validations can reference
other objects, we cannot be sure the evaluation is possible, and they
must be subject to the same rules as all temporary values and be skipped
when not required in the destroy graph.