allow unknown error_message in validate
When evaluating variable validation, we may have valid error_message
values which are unknown due to input other referenced input variables
always being unknown during the validate walk. In the rare case where
the variable validation is known but the error_message is not, we need
to be able to proceed to the plan phase in order to check the validation
more thoroughly.