stacks: support sensitive input values in planned changes (#35640)
* stacks: support sensitive input values in planned changes

* stacks: Fix panic when applying sensitive inputs

The stacks-specific check that a given plan's input variable values
remain unchanged at apply time would previously panic when given a
marked value. Marks are not important for this check, so we simply
remove them to ensure that it passes. This seems in line with how
AssertObjectCompatible works, which I'm taking as precedent.

The new test added in this commit panics without the corresponding code
change.

---------

Co-authored-by: Alisdair McDiarmid <alisdair@users.noreply.github.com>