fix connections for ephemeral closer
The graph building around nodeEphemeralResourceClose was not correct,
and allowed the provider to be shut down before the ephemeral resource
was closed.

The nodeEphemeralResourceClose is not connected to the main provider
node because it's not added until after the providers are all connected.
Because the close node needs to wait on all dependencies of the
ephemeral resource, the usual graph building process no longer works.
There have been a lot of improvements since the original provider close
transformer was introduced, so we should be able to refactor that in
terms of what consumes the provider rather than just looking for edges
from the provider node.