make ephemeral resource work during destroy
Ephemeral resource behavior is a strange mix of temporary values and
existing resource types, and their operation during destroy is going to
be more difficult to manage than normal resources.

The pruneUnusedNodesTransformer is probably no longer equipped to handle
the situation, and will probably need to be build in a different manner,
but this minor change allows us to move forward for  now.