connect ephemeral close nodes more accurately
In order to find the correct location to connect the ephemeral close
nodes, we need to walk descendants if the ephemeral value until we reach
the first node which nothing references. That should signal the end of
an evaluation chain, and can be a dependency for closing the ephemeral
value.