refactor and simplify pruneUnusedNodesTransformer
With the missing destroy node types sorted out, this transformer can now
be designed around finding the nodes we need to keep for destroy
operations, rather than the whack-a-mole game of detecting nodes we
don't want to keep.

When executing a destroy operation, we can never be certain that the
state will contain all the data required to evaluate the entire
configuration. This could be because of a previously failed apply
operation, or external changes made out of band. This means that once we
process the configuration to get the providers hooked up, we need to
remove any nodes that correspond to configuration values which may not
be available to evaluate.

In a destroy operation we primarily need three categories of nodes to
complete the plan and apply:
 - Nodes corresponding to values in the state
 - Providers required to destroy those values obtained from the state
 - Any node which a provider depends on, to ensure its configuration can
   be fully evaluated

Other than that minimal set of nodes, the rest are no longer necessary
and should be skipped to avoid errors which could block the apply
operation.