don't error on eval of missing instances
While in most cases a missing instance during evaluation would be a bug
in terraform, the insertion of postconditions into instance planning has
created a situation where not all instances will be available when
GetResource is called. We do have a special context for evaluating self
alone, but postconditions also need to have the module context available
too.

We may be able to come up with some other way to create an evaluation
context which handles self more selectively, but that looks like a large
change for a situation which should not otherwise happen. If this were
to mask a legitimate error, the fact that an unexpected DyanamicVal were
encountered will raise the problem again, and we can correlate that with
the logged warning.