add discrete codepaths for ephemeral apply nodes
It turns out the reuse of the plan nodes for ephemerals during apply was
interfering with how checks are handled in the plan. Since checks
already have an established pattern of use, we can avoid the whole issue
by using only the resource apply nodes to kick of ephemeral values.

This copies the subset of expansion logic needed from the resource plan
node into the apply node. Since the apply time evaluation is much
simpler, duplicated code is not overly complex, but we may want to
revisit this for refactoring in the future.