create a ui hook for unknown ephemeral config
This is a stopgap to give users some feedback when an ephemeral resource
contains unknowns and cannot be opened.

There is no json hook implementation so that the machine-readable spec
is not bound to the new output until further review. It doesn't really
fit with the current json ui model to have messages about what isn't
going to happen.