remove AssertObjectCompatible from refresh
AssertObjectCompatible can be expensive, and the log messages at this
point are no longer needed.

We can't assert anything during refresh, because the provider can always
be correcting for external drift here. The call and logs ware originally
kept in the code to help with legacy SDK provider updates, and have not
been especially useful in a long time.