use a set to remove duplicates
Due to the need to search through parent directory depends_on for data
sources, the dependencies listed for data sources can contain numerous
duplicates. In the worst-case configurations, with nested modules that
have multiple depends_on entries, duplicates can be listed hundreds of
times.

While `dag.Set` is not a great generic data structure, use that for
de-duplication because it's the same structure used internally by
`dag.Graph` so we can be sure nodes are consistently unique.