simplify cycle check for inter-provider deps
The design of Terraform inherently can create cycles for replacement
across providers, and the interim solution was to simply check for any
cycles to skip these dependencies. Checking the entire graph for each
connection was a bit silly, but doesn't have any problems for most
configurations. This is quadratic though, so extremely large numbers of
interdependent replacement ops may slow down to the point of making a
plan impossible.

While we don't have a solution for configuring the provider out of band,
or quickly checking the provider dependencies as noted in the FIXME, we
can do nearly the same thing just by making sure our `from` node isn't
already an ancestor of the `to`, essentially preemptively checking for
the cycle.