encoding of changes should always return a value
Terraform always attempts to generate a plan even when encountering an
error, so we need to ensure that all plan elements have valid values to
store and inspect later. Ensuring that `Changes.Encode` always returns a
valid value avoids the need for nil checks elsewhere while returning an
error to the user.