resources must be refreshed before plan
The order of operations was incorrect here, with the refresh happening
after the plan was already created. Invert the calls so the updated
state is reflected in the plan output.

We don't need to carry forward the defer response from refresh, since in
the rare case that a read does need to be deferred, the
PlanResourceChange call will return the same response.