WriteOnlyPaths was returning the wrong paths
WriteOnlyPaths was returning some paths from values which did not exist,
while missing others which did. The process is exactly the same as for
SensitivePaths, so replace the methods with an exact copy changing out
the appropriate field and method names. There's probably a way to
abstract these into a single method, but it wasn't apparent if or how
that's possible without a much more complex method, so for now I left
them as duplicates with comments to check the other when changes are
made.