Temporally disable new windows symlink behavior
EvalSymlinks and Readlink on Windows used to return incorrect
information for windows junctions which were specifically not symlinks
prior to go1.23. Unfortunately Terraform uses `EvalSymlinks` in
situations where Windows users had come to expect their configuration to
work as if it were symlinked even though it used a different filesystem
structure. Set winsymlink=0 to restore the existing behavior until
we can determine the best course of action for Windows users.