Remove irrelevant link to issue.
The docs link to https://github.com/hashicorp/terraform/issues/13549 which was closed. Gathering anything relevant or of use from that ~100 comment long issue thread is not a particularly useful endeavor, so given that the information is true and not likely to be changed, simply removing the issue link and leaving the warning.