Backend/azure: `subscription_id` infer from Azure CLI & a way to skip *unnecessary* management plane API call (#36623) (#36677)
* backend/azure: Infer `subscription_id` from Azure CLI if not specified

* Avoid management plane call if not necessary

There are four auth scenarios of the blob/container:
1. User specifies the SAS token
2. User specifies the shared access key
3. User specified to use AAD auth (and credential provided)
4. None of the above, management plane API call needed to list the shared access key and use it to auth

For 1, 2 and 3, the management plane API can be skipped in most of the cases, except the target storage account is using private DNS zone. The blob/container data plane client requires the user to specify the base URI, which is not deterministic if the storage account opt in the private DNS zone. In this case, an additional management plane `GET` is required against the storage account, to retrieve the blob endpoint. While if private DNS zone is not used, the base URI can be composed in a fixed pattern, with the storage account name and container&blob name. Hence no management plane API call is needed.

The user is expected to use the `subscription_id` and `resource_group_name` to indicate the above intent: Only if both are specified, the additional `GET` call will be invoked to get the accurate blob endpoint.

(NOTE: the `subscription_id` can be inferred from the Azure CLI if unspecified)

* No need for AAD auth if shared key or sas is specified

* update

* Add test

* update comments

* changelog

Co-authored-by: magodo <wztwcy@gmail.com>