hcl diagnostics must check Subject source
The source check was skipped if either value was nil, but we can't be
sure they are equal if there is no value. This may prove to be
problematic is tests, as diagnostic equality for deduplication isn't
strict equality, but rather relies on being able to prove the context is
also equal.