Decode action blocks and action_triggers inside resource blocks. (#37030)
* Decode action blocks and action_triggers inside resource blocks.

This commit adds decoding of action and action_triggers inside terraform configuration. I added an Actions experiment as a hacky way of keeping the functionality out of main until we're ready for the alpha; this may never be an experiment but it's a handy feature flag so we don't have to do all the work in a long-lived feature branch.

* remove legacy shim handling
* validate that the referenced entry in actions is indeed an action