reduce changes lookups in GetResource
Expanded resources always require fetching all instances for evaluation,
but the changes data structure requires iterating over all instances for
each call to GetResourceInstanceChange, so it turns out to be an n^2
operation. This wasn't very visible in most configs with relatively few
numbers of instances, but when you have a large number of instances
referencing another resource with a large number of instances (e.g. the
second resource uses the first as the for_each argument), you end
up with (n^2)^2 lookups. Because the instance changes are in a global
structure, this single point of lock contention would hold up the entire
plan.