allow outputs without top-level ephemeral marks
There is currently no way to construct an output value which does not
come directly from an ephemeral resource or input. Even if the
constructed value is entirely composed of ephemeral attributes or
elements, because there is no action to manually mark something as
ephemeral (other than a variable block), the resulting value itself is
not marked as ephemeral, and fails validation.

Instead of strictly checking for an ephemeral mark on the output value,
we can accept any value because the result will always be entirely
ephemeral anyway. This mirrors the variable block behavior, which can
accept any type of value as input, but is evaluated as ephemeral.