actions: implement expansion and planning actions
This implementation is very barebones, but builds the skelleton for
all upcoming features in actions. It includes the graph implementation
for planning actions including the expansion of actions and dealing
with multiple action triggers.

We plan all actions as part of their triggering resource for now.
All action_triggers depend on the triggering resources plan to
determine if an action should run or not, therefore we can plan
the actions as part of the triggering resource.

This implementation ignores conditions, deferrals, provider arguments
and it does not add the actions to the planfile; mostly to enable more
people to work on these issues in parallel.