refactor config generation
Config generation was mixing the "what" to generate with the "how" to
generate it, making it impossible to use other means of creating a
configuration value. We extract the filtering of the state into a
configuration value from the formatting code, maintaining the quirks
necessary to work with the legacy SDK.

We also removed all marks handling from the config generation, because
non-schema marks can only come from configuration, which this value is
not a part of.