get correct schema when generating query config
During the latest refactor, the wrong schema was fetched when generating
configuration from a list resource. Having the list resource schema
instead of the managed resource schema prevented the config generation
from matching the computed attributes within the state, causing them to
show up in the configuration.

It turns out the tests for this didn't have computed attributes, so the
accidental swap during refactoring didn't cause the tests to fail. A
small addition to the tests adding an `id` attribute picks the
regression.