Fix `forceReplace` modification race
Terraform stores `-replace` addresses in a slice that is being passed by
reference to graph nodes. When it encounters a `replace_triggered_by`,
it adds the address of the resource that triggered the replacement.
The problem is that all graph nodes reuse the same slice. This commit
fixes this problem by making a copy of the slice.