terraform_data sensitive and write-only tools
Add new objects to the terraform_data resource to handle sensitive and
write-only attributes.

The sensitive object works just like the top-level input+output
attributes, except they are always marked as sensitive.

The write_only object also has input and output values, which work much
like the top-level input and output. The difference is that the
write_only.input is truly "write only" and can accept ephemeral values.
There is also a `version`, which allows the user to determine exactly
when the ephemeral value will be stored.

The sensitive and write_only objects also have a "replace" attribute,
allowing changes in the output to work just like "triggers_replace". I
opted to use replace rather than triggers_replace, since the
former is a bool flag, while the latter is tracking an actual value.